+++ packages/social-controllers/src/SocialController-method-action-types.ts
@@ -56,6 +56,18 @@ export type SocialControllerUpdateFollowingAction = {
   handler: SocialController['updateFollowing'];
 };
 
+/** Opts the current user out of the PnL leaderboard. */
+export type SocialControllerOptOutOfLeaderboardAction = {
+  type: `SocialController:optOutOfLeaderboard`;
+  handler: SocialController['optOutOfLeaderboard'];
+};
+
+/** Opts the current user back into the PnL leaderboard. */
+export type SocialControllerOptInToLeaderboardAction = {
+  type: `SocialController:optInToLeaderboard`;
+  handler: SocialController['optInToLeaderboard'];
+};
+
 /**
  * Union of all SocialController action types.
  */
@@ -63,4 +75,6 @@ export type SocialControllerMethodActions =
   | SocialControllerUpdateLeaderboardAction
   | SocialControllerFollowTraderAction
   | SocialControllerUnfollowTraderAction
-  | SocialControllerUpdateFollowingAction;
+  | SocialControllerUpdateFollowingAction
+  | SocialControllerOptOutOfLeaderboardAction
+  | SocialControllerOptInToLeaderboardAction;