+++ packages/social-controllers/src/SocialService-method-action-types.ts
@@ -141,6 +141,18 @@ export type SocialServiceUnfollowAction = {
   handler: SocialService['unfollow'];
 };
 
+/** Opts the current user out of the PnL leaderboard. */
+export type SocialServiceOptOutOfLeaderboardAction = {
+  type: `SocialService:optOutOfLeaderboard`;
+  handler: SocialService['optOutOfLeaderboard'];
+};
+
+/** Opts the current user back into the PnL leaderboard. */
+export type SocialServiceOptInToLeaderboardAction = {
+  type: `SocialService:optInToLeaderboard`;
+  handler: SocialService['optInToLeaderboard'];
+};
+
 /**
  * Union of all SocialService action types.
  */
@@ -153,4 +165,6 @@ export type SocialServiceMethodActions =
   | SocialServiceFetchPositionByIdAction
   | SocialServiceFetchFollowingAction
   | SocialServiceFollowAction
-  | SocialServiceUnfollowAction;
+  | SocialServiceUnfollowAction
+  | SocialServiceOptOutOfLeaderboardAction
+  | SocialServiceOptInToLeaderboardAction;